requirements.txt
requests
pandas
BeautifulSoup4
python-dotenv
